Simplify driver interfaces
As part of improvement of driver interfaces, the
public method execute statement is being removed.
Also list_tables() will no longer be a part of the
session interface and is being moved to driver
interface. The non-pooled qldb driver implementation
is also being removed because it did not appear to
serve much purpose and does not really align with a
practical use case.