Rename PooledQldbDriver
Going forward there will be one
driver which would behave like the
PooledQldbDriver. There wasn't much sense
in naming it as PooledQldbDriver so we
decided to change it to QldbDriver whihc
is cleaner. The same applied to
PooledQldbSession. Also since we are
completely removing non pooled implementations,
there isn't much of value in keeping an
abstract class for a single driver implementation,
so we decided to remove them as well.