Improve session pooling algorithm
Earlier, while executing the Transaction, if there
was an InvalidSessionException thrown from the service,
the driver would send another request to create a new
session and retry the Tx with the new session.
This mechanism did not take advantage of a possibly
healthy session in the pool and maybe we could have
avoided creating new session. This change fixes that
and now, whenever ther eis ISE, the driver will try
with the next session in the pool. If all the
sessions in the pool are exhausted, only then the
driver will create a new session.