Remove retries for an expired transaction
Earlier a transaction used to be retried
infinite number of times. We don't see value
in retrying a transaction as it is which outlived
the regular transaction life. With this change, a
transaction will not be retried if it failed due to
a transaction expiry and the execption will be bubbled
all the way to the application code using the driver.